Update to be the right state of the work. Too many branches.
