+++ jujugui/static/gui/src/app/init.js
@@ -10,6 +10,8 @@ const hotkeys = require('./init/hotkeys');
 const csUser = require('./init/charmstore-user');
 const cookieUtil = require('./init/cookie-util');
 
+const newBakery = require('./utils/bakery-utils');
+
 const ComponentRenderersMixin = require('./init/component-renderers-mixin');
 const DeployerMixin = require('./init/deployer-mixin');
 
@@ -118,33 +120,12 @@ class GUIApp {
     };
 
     /**
-      Shows the login notification
-      @param {Object} content JSX of the content.
-    */
-    const showLoginNotification = (content) => {
-      const holder = document.getElementById('login-notification');
-      let dismiss = null;
-      if (stateGetter().root !== 'login') {
-        dismiss = () => {
-          ReactDOM.unmountComponentAtNode(holder);
-        };
-      }
-      ReactDOM.render(
-        <Notification
-          content={content}
-          dismiss={dismiss}
-          extraClasses="four-col"
-          isBlocking={true}
-        />, holder);
-    };
-    /**
       A bakery instance.
       Used to perform requests on a macaroon authenticated endpoints.
       @type {Object}
     */
-    this.bakery = yui.juju.bakeryutils.newBakery(
-      config, this.user, stateGetter, cookieSetter, webHandler,
-      showLoginNotification);
+    this.bakery = newBakery(
+      config, this.user, stateGetter, cookieSetter, webHandler);
     /**
       A charm store API client instance.
       Used to retrieve information about charms and bundles via the charm store.
