+++ jujugui/templates/index.html.mako
@@ -360,7 +360,6 @@ with this program.  If not, see <http://www.gnu.org/licenses/>.
             'juju-env-web-handler',
             'juju-models',
             'jujulib-utils',
-            'bakery-utils',
             'net-utils',
             // juju-views group
             'd3-components',
